Fix some compiler warnings

+++ MonoPitchHMM.h
@@ -28,13 +28,11 @@ class MonoPitchHMM : public SparseHMM
 public:
     MonoPitchHMM();
     const std::vector<double> calculateObsProb(const vector<pair<double, double> >);
-    // double getMidiPitch(size_t index);
-    // double getFrequency(size_t index);
     void build();
     double m_minFreq; // 82.40689f/2
-    size_t m_nBPS;
-    size_t m_nPitch;
-    size_t m_transitionWidth;
+    int m_nBPS;
+    int m_nPitch;
+    int m_transitionWidth;
     double m_selfTrans;
     double m_yinTrust;
     vector<double> m_freqs;
